+++ android/.idea/gradle.xml
@@ -36,7 +36,7 @@
             <option value="$PROJECT_DIR$/core/network" />
             <option value="$PROJECT_DIR$/feature" />
             <option value="$PROJECT_DIR$/feature/account" />
-            <option value="$PROJECT_DIR$/feature/analyze" />
+            <option value="$PROJECT_DIR$/feature/analysis" />
             <option value="$PROJECT_DIR$/feature/auth" />
             <option value="$PROJECT_DIR$/feature/dictionary" />
             <option value="$PROJECT_DIR$/feature/review" />
