+++ android/settings.gradle
@@ -35,7 +35,7 @@ include ':core:designsystem'
 include ':core:database'
 include ':feature:auth'
 include ':feature:setting'
-include ':feature:analyze'
+include ':feature:analysis'
 include ':feature:dictionary'
 include ':feature:review'
 include ':feature:account'
